refactor: moved test helper into subfolder
